ci: zero-warning clippy gate (Epic 12 retro AI-12.1)
  Replace the diff-based "0-new" clippy gate with an absolute `-D warnings`
  gate so the lint baseline can no longer rot silently — it had reached 257
  (--all-targets) warnings while every story honestly reported "0-new".

  - Baseline 257 -> 0 across default / telegram+cron / vector-search. Real
    fixes for production idiom lints; documented per-lint allows in
    lib.rs+main.rs (extends the existing block) + a package-wide Cargo.toml
    [lints] table (covers test crates + build.rs); module allows on shared
    test helpers.
  - ci.yml: warn-only + merge-base "0-new" diff -> `cargo clippy
    --all-targets -- -D warnings` (default + telegram/cron per-PR;
    vector-search lane added to nightly-vector-search.yml, since its ort
    build needs network). Toolchain pinned @stable -> @1.94 for lint
    reproducibility.
  - Fix latent bug surfaced by the sweep: server.rs dropped two bounded
    mpsc send futures (`let _ = tx.send(..)` with no await) on the 12.2d
    consolidation path, so the card/notice was never sent. Now awaited.
  - MSRV: rust-version 1.85 -> 1.91 (true floor; floor_char_boundary is
    stable 1.91 and already used). Fixes 2 manual_is_multiple_of sites and
    a build.rs nested-if -> let-chain.
  - AI-12.4: migrate 3 vector-search tests off deprecated
    RedactionRecord::forget -> redact(key, "", ts) (byte-identical record).

  Verification: 3972 tests pass / 0 fail; all 3 clippy lanes green at
  -D warnings; fmt clean.

  Also carries pre-existing test-infra tweaks: tests_tui/harness.py
  RUST_LOG=debug and ci.yml pytest -n auto.

+++ perf-tests/subagent_spawn.rs
@@ -199,7 +199,7 @@ async fn memory_rss_per_agent() {
         .unwrap()
         .resident;
     let page_size_kb = unsafe { libc::sysconf(libc::_SC_PAGESIZE) as u64 / 1024 };
-    let delta_kb = (after - before) as u64 * page_size_kb;
+    let delta_kb = (after - before) * page_size_kb;
     println!(
         "RSS delta for 10 agents: {} KB ({} KB/agent)",
         delta_kb,
